rename: 인터셉터 분리 및 인스턴스 폴더 구조 변경
